Разделение представления на вкладки и вью модели на соответствующие вкладки
