+++ DOfficeCore/ViewModels/MainWindowViewModel.cs
@@ -6,6 +6,7 @@ using DOfficeCore.Services.Interfaces;
 using DOfficeCore.ViewModels.Core;
 using System;
 using System.Collections.ObjectModel;
+using System.Windows.Input;
 
 namespace DOfficeCore.ViewModels
 {
@@ -60,7 +61,60 @@ namespace DOfficeCore.ViewModels
             #endregion
         }
 
-        #region Свойства окна дневника
+        #region Заголовок окна
+        /// <summary>Заголовок окна</summary>
+        private string _Title = "Кабинет врача";
+        /// <summary>Заголовок окна</summary>
+        public string Title
+        {
+            get => _Title;
+            set => Set(ref _Title, value);
+        }
+        #endregion
+
+        #region Команды
+
+        #region Команда закрытия программы
+        /// <summary>Команда закрытия программы</summary>
+        public ICommand ClosingAppCommand { get; }
+        /// <summary>Команда закрытия программы</summary>
+        private void OnClosingAppCommandExecuted(object parameter)
+        {
+            _Logger.WriteLog("EXIT", "Закрытие программы.");
+        }
+
+        private bool CanClosingAppCommandExecute(object parameter) => true;
+
+        #endregion
+
+        #region Команда загрузки данных
+        /// <summary>Команда Загрузки данных</summary>
+        public ICommand LoadDataCommand { get; }
+        /// <summary>Команда Загрузки данных</summary>
+        private void OnLoadDataCommandExecuted(object parameter)
+        {
+            _Logger.WriteLog("INFO");
+
+            _ViewCollection.DataCollection = _DataProviderService.LoadDataFromFile("file.json");
+            _ViewCollectionProvider.DiagnosisFromDataToView();
+            var temp = _DataProviderService.LoadDoctorsFromFile("Doctors.json");
+            if (temp != null) Doctors = new ObservableCollection<string>(temp);
+            temp = _DataProviderService.LoadDoctorsFromFile("Position.json");
+            if (temp != null) Position = new ObservableCollection<string>(temp);
+
+            _Logger.WriteLog("DONE");
+        }
+
+        private bool CanLoadDataCommandExecute(object parameter) => true;
+        #endregion
+
+        #endregion
+
+        #region Сервисы
+
+        #region ViewCollection : IViewCollection - Коллекция данных
+        private readonly IViewCollection _ViewCollection;
+        #endregion
 
         #region Сервис обработки строк
         private readonly ILineEditorService _LineEditorService;
@@ -82,176 +136,10 @@ namespace DOfficeCore.ViewModels
         private readonly IDiaryBoxProvider _DiaryBoxProvider;
         #endregion
 
-        #region Заголовок окна
-        /// <summary>Заголовок окна</summary>
-        private string _Title = "Кабинет врача";
-        /// <summary>Заголовок окна</summary>
-        public string Title
-        {
-            get => _Title;
-            set => Set(ref _Title, value);
-        }
         #endregion
+
         
-        #region ViewCollection : IViewCollection - Коллекция данных
-        private readonly IViewCollection _ViewCollection;
-        public IViewCollection ViewCollection { get => _ViewCollection; }
-        #endregion
 
-        #region EnableTextBox : bool - Состояние возможности редактирования текстового окна
-        /// <summary>Состояние возможности редактирования текстового окна</summary>
-        private bool _EnableTextBox = true;
-        /// <summary>Состояние возможности редактирования текстового окна</summary>
-        public bool EnableTextBox
-        {
-            get => _EnableTextBox;
-            set => Set(ref _EnableTextBox, value);
-        }
-        #endregion
-
-        #region FocusedDataGrid : DataGrid - Имя датагрида, который сейчас находится в фокусе
-
-        /// <summary>Имя датагрида, который сейчас находится в фокусе</summary>
-        private string _FocusedDataGrid;
-
-        /// <summary>Имя датагрида, который сейчас находится в фокусе</summary>
-        public string FocusedDataGrid
-        {
-            get => _FocusedDataGrid;
-            set => Set(ref _FocusedDataGrid, value);
-        }
-
-        #endregion
-
-        #region MultiBox : string - Содержимое мультибокса
-
-        /// <summary>Содержимое мультибокса</summary>
-        private string _MultiBox;
-
-        /// <summary>Содержимое мультибокса</summary>
-        public string MultiBox
-        {
-            get => _MultiBox;
-            set => Set(ref _MultiBox, value);
-        }
-
-        #endregion
-
-        #region DiaryBox : string - Содержимое дневника
-
-        /// <summary>Содержимое дневника</summary>
-        private string _DiaryBox;
-
-        /// <summary>Содержимое дневника</summary>
-        public string DiaryBox
-        {
-            get => _DiaryBox;
-            set => Set(ref _DiaryBox, value);
-        }
-
-        #endregion
-
-        #region ChoosenDate : Datetime - Выбранная дата
-
-        /// <summary>Выбранная дата</summary>
-        private DateTime _ChoosenDate = DateTime.Now;
-
-        /// <summary>Выбранная дата</summary>
-        public DateTime ChoosenDate
-        {
-            get => _ChoosenDate;
-            set => Set(ref _ChoosenDate, value);
-        }
-
-        #endregion
-
-        #region Doctors : ObservableCollection<string> - Список докторов
-
-        /// <summary>Список докторов</summary>
-        private ObservableCollection<string> _Doctors;
-
-        /// <summary>Список докторов</summary>
-        public ObservableCollection<string> Doctors
-        {
-            get => _Doctors;
-            set => Set(ref _Doctors, value);
-        }
-
-        #endregion
-
-        #region Position : ObservableCollection<string> - Должность
-
-        /// <summary>DESCRIPTION</summary>
-        private ObservableCollection<string> _Position;
-
-        /// <summary>DESCRIPTION</summary>
-        public ObservableCollection<string> Position
-        {
-            get => _Position;
-            set => Set(ref _Position, value);
-        }
-
-        #endregion
-
-        #region CurrentPosition : string - Поле ввода для должностей
-
-        /// <summary>Поле ввода для должностей</summary>
-        private string _CurrentPosition;
-
-        /// <summary>Поле ввода для должностей</summary>
-        public string CurrentPosition
-        {
-            get => _CurrentPosition;
-            set => Set(ref _CurrentPosition, value);
-        }
-
-        #endregion
-
-        #region CurrentDoctor : string - Поле ввода для докторов
-
-        /// <summary>Поле ввода для докторов</summary>
-        private string _CurrentDoctor;
-
-        /// <summary>Поле ввода для докторов</summary>
-        public string CurrentDoctor
-        {
-            get => _CurrentDoctor;
-            set => Set(ref _CurrentDoctor, value);
-        }
-
-        #endregion
-
-        #endregion
-
-        #region Свойства окна обработчика строк
-
-        #region TextForEditing : string - Текст, который необходимо обработать
-
-        /// <summary>Текст, который необходимо обработать</summary>
-        private string _TextForEditing;
-
-        /// <summary>Текст, который необходимо обработать</summary>
-        public string TextForEditing
-        {
-            get => _TextForEditing;
-            set => Set(ref _TextForEditing, value);
-        }
-
-        #endregion
-
-        #region RawLines : ObservableCollection<string> - Необработанная коллекция строк
-
-        /// <summary>Необработанная коллекция строк</summary>
-        private ObservableCollection<string> _RawLines;
-
-        /// <summary>Необработанная коллекция строк</summary>
-        public ObservableCollection<string> RawLines
-        {
-            get => _RawLines;
-            set => Set(ref _RawLines, value);
-        }
-
-        #endregion
-        #endregion
+        
     }
 }
