+++ DOfficeCore/Views/LinesEditView.xaml.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace DOfficeCore.Views
+{
+    /// <summary>
+    /// Логика взаимодействия для LinesEditView.xaml
+    /// </summary>
+    public partial class LinesEditView : UserControl
+    {
+        public LinesEditView()
+        {
+            InitializeComponent();
+        }
+    }
+}
